chore: Refactor periodic requests initialization
The code changes refactor the initialization of periodic requests in the `working_state.c` file. The `periodic_requests` function now calls `ccp_message_sender_send_request` with the `CCP_AT_TM` parameter. The `init_periodic_requests` function is updated to use `periodic_task_init_b` with a delay of 10000 milliseconds.

This commit message follows the established convention of using a verb in the imperative form to describe the purpose of the changes.